streamserver/server/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
gunicorn
psutil
yt-dlp
